+++ ScrumPoker-react/Models/PlayerModel.cs
@@ -4,7 +4,7 @@ public class PlayerModel
 {
     public string Id { get; set; } = "";
     public string Name { get; set; } = "";
-    public int Vote { get; set; }
+    public float Vote { get; set; }
     public string Mode { get; set; } = "";
 
 }
